Updated migration+models

+++ database/migrations/2025_12_08_093501_users.php
@@ -15,7 +15,7 @@ return new class extends Migration {
               $table->id();
               
      $table->unsignedBigInteger('company_id')->nullable(); 
-            $table->integer('type_id');
+            $table->unsignedBigInteger('type_id');
             $table->string('name');
             $table->string('email')->unique();
             $table->string('password');
